moved User and Task under data subpackage
