+++ src/e2e/java/edu/mikedev/task_manager/BDDSteps.java
@@ -1,6 +1,7 @@
 package edu.mikedev.task_manager;
 
 import edu.mikedev.task_manager.controller.TaskManagerController;
+import edu.mikedev.task_manager.data.Task;
 import edu.mikedev.task_manager.model.DBLayer;
 import edu.mikedev.task_manager.model.HibernateModel;
 import edu.mikedev.task_manager.model.Model;
