+++ src/it/java/edu/mikedev/task_manager/ModelIT.java
@@ -1,5 +1,7 @@
 package edu.mikedev.task_manager;
 
+import edu.mikedev.task_manager.data.Task;
+import edu.mikedev.task_manager.data.User;
 import edu.mikedev.task_manager.model.DBLayer;
 import edu.mikedev.task_manager.model.HibernateModel;
 import edu.mikedev.task_manager.utils.HibernateDBUtils;
