+++ src/main/java/edu/mikedev/task_manager/controller/LoginPageController.java
@@ -1,6 +1,6 @@
 package edu.mikedev.task_manager.controller;
 
-import edu.mikedev.task_manager.Task;
+import edu.mikedev.task_manager.data.Task;
 import edu.mikedev.task_manager.ui.LoginPage;
 import edu.mikedev.task_manager.ui.RegistrationPage;
 import edu.mikedev.task_manager.ui.UserTasksPage;
