+++ src/main/java/edu/mikedev/task_manager/data/User.java
@@ -1,4 +1,4 @@
-package edu.mikedev.task_manager;
+package edu.mikedev.task_manager.data;
 
 import java.util.Objects;
 import java.util.Set;
