+++ src/main/java/edu/mikedev/task_manager/model/HibernateDBLayer.java
@@ -1,7 +1,7 @@
 package edu.mikedev.task_manager.model;
 
-import edu.mikedev.task_manager.Task;
-import edu.mikedev.task_manager.User;
+import edu.mikedev.task_manager.data.Task;
+import edu.mikedev.task_manager.data.User;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
