+++ src/main/resources/task.hbm.xml
@@ -1,4 +1,4 @@
-<hibernate-mapping package="edu.mikedev.task_manager">
+<hibernate-mapping package="edu.mikedev.task_manager.data">
     <!-- class name is the fully qualified name of the POJO class -->
     <!-- table is the name of the database table which represents the Student -->
     <class name="Task" table="tasks">
