+++ src/test/java/edu/mikedev/task_manager/controller/TestUpdateTaskBindings.java
@@ -1,7 +1,7 @@
 package edu.mikedev.task_manager.controller;
 
-import edu.mikedev.task_manager.Task;
-import edu.mikedev.task_manager.User;
+import edu.mikedev.task_manager.data.Task;
+import edu.mikedev.task_manager.data.User;
 import edu.mikedev.task_manager.model.Model;
 import edu.mikedev.task_manager.utils.UIScenarios;
 import org.assertj.swing.annotation.GUITest;
@@ -12,7 +12,6 @@ import org.assertj.swing.fixture.JLabelFixture;
 import org.assertj.swing.junit.runner.GUITestRunner;
 import org.assertj.swing.junit.testcase.AssertJSwingJUnitTestCase;
 import org.assertj.swing.util.Triple;
-import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
